final project genAi/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
langchain==0.0.348
langchain-community==0.0.29
langgraph==0.0.40
chromadb==0.4.15
duckduckgo-search==3.9.6
sentence-transformers==2.2.2
transformers==4.35.0
torch>=1.12.0
python-dotenv==1.0.0
pydantic==2.5.3
tavily-python==0.3.1
google-generativeai==0.3.2
google-ai-generativelanguage==0.4.0
langchain-core==0.1.33
langsmith==0.0.87
langchain-google-genai==0.0.5
langchain-text-splitters==0.0.1
protobuf==4.25.8
pypdf==6.1.0